test(store): move the failing-client fake into a shared fakes file
