+++ internal/store/postgres/org_billing_repository_test.go
@@ -2,13 +2,8 @@ package postgres
 
 import (
 	"context"
-	"database/sql"
-	"database/sql/driver"
-	"errors"
 	"testing"
 
-	"github.com/jmoiron/sqlx"
-	"github.com/raystack/frontier/pkg/db"
 	"github.com/raystack/salt/rql"
 	"github.com/stretchr/testify/assert"
 )
@@ -222,20 +217,6 @@ func TestPrepareGroupByQuery(t *testing.T) {
 		})
 	}
 }
-
-// txnFailClient returns a client whose connections always fail, so any
-// transaction begin returns an error.
-func txnFailClient() *db.Client {
-	return &db.Client{DB: sqlx.NewDb(sql.OpenDB(failConnector{}), "postgres")}
-}
-
-type failConnector struct{}
-
-func (failConnector) Connect(context.Context) (driver.Conn, error) {
-	return nil, errors.New("connection failed")
-}
-
-func (failConnector) Driver() driver.Driver { return nil }
 
 func TestOrgBillingRepository_Search(t *testing.T) {
 	t.Run("should return error when the transaction cannot start", func(t *testing.T) {
